project configuration refactored
Switched old icons to another svg's, rejected debugbar, .env.example refactored, some code refactored, added images to git, refactored code into OrderController (bug with commas in total price), refactored code in update methods for cigarette form and liquid form (rejected require rule from validation), added styles to error messages for liquid-edit form

+++ database/migrations/2024_12_04_164001_create_orders_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('city');
             $table->string('post_office');
             $table->string('details', 191)->nullable();
-            $table->float('total');
+            $table->decimal('total');
             $table->string('payment_method');
             $table->json('products');
             $table->string('status')->default('in processing');
